Move specifics into oscillo project
